+++ src/ducks/Studio/Chart/Sidepanel/HolderDistribution/index.module.scss
