+++ src/ducks/Studio/Chart/Sidepanel/MetricsExplanation/DataInfo.js
