+++ src/ducks/Studio/Chart/Sidepanel/MetricsExplanation/Explanations.module.scss
