+++ src/ducks/Studio/Chart/Sidepanel/MetricsExplanation/index.module.scss
