+++ src/ducks/Studio/Chart/Sidepanel/index.js
