+++ src/ducks/Studio/Chart/Sidepanel/index.module.scss
