+++ src/ducks/Studio/Chart/Sidepanel/panes.js
