+++ src/ducks/Studio/Header/Settings.js
@@ -4,9 +4,9 @@ import { withRouter } from 'react-router-dom'
 import Button from '@santiment-network/ui/Button'
 import Icon from '@santiment-network/ui/Icon'
 import AdvancedCalendar from '../../../components/AdvancedCalendar'
-import MetricsExplanation from '../Chart/Sidepane/MetricsExplanation'
+import MetricsExplanation from '../Chart/Sidepanel/MetricsExplanation'
+import { METRICS_EXPLANATION_PANE } from '../Chart/Sidepanel/panes'
 import { getIntervalByTimeRange } from '../../../utils/dates'
-import { METRICS_EXPLANATION_PANE } from '../../Studio/Chart/Sidepane/panes'
 import styles from './Settings.module.scss'
 import ShareModalTrigger from '../../../components/Share/ShareModalTrigger'
