+++ src/ducks/Studio/Sidebar/MetricSelector/index.js
@@ -1,6 +1,6 @@
 import React, { useMemo } from 'react'
 import Category from './Category'
-import { TopHolderMetric } from '../../Chart/Sidepane/HolderDistribution/metrics'
+import { TopHolderMetric } from '../../Chart/Sidepanel/HolderDistribution/metrics'
 import styles from './index.module.scss'
 
 const MetricSelector = ({ categories = {}, availableMetrics, ...rest }) => {
