+++ src/ducks/Studio/index.js
@@ -1,315 +1,190 @@
 import React, { useState, useEffect } from 'react'
-import cx from 'classnames'
-import StudioSidebar from './Sidebar'
-import StudioMain from './Main'
-import {
-  DEFAULT_SETTINGS,
-  DEFAULT_OPTIONS,
-  DEFAULT_METRICS,
-  DEFAULT_METRIC_SETTINGS_MAP
-} from './defaults'
-import { MAX_METRICS_AMOUNT } from './constraints'
-import { generateShareLink } from './url'
-import { trackMetricState } from './analytics'
-import { useTimeseries } from './timeseries/hooks'
-import { buildAnomalies } from './timeseries/anomalies'
-import { buildComparedMetric } from './Compare/utils'
-import { TOP_HOLDERS_PANE } from './Chart/Sidepane/panes'
-import { useClosestValueData } from '../Chart/hooks'
+import Sidebar from './Sidebar'
+import { newChartWidget } from '../Studio2/Widget/ChartWidget'
+import { newHolderDistributionWidget } from '../Studio2/Widget/HolderDistributionWidget'
+import SelectionOverview from '../Studio2/SelectionOverview'
+import Main from '../Studio2/Main'
 import { getNewInterval, INTERVAL_ALIAS } from '../SANCharts/IntervalSelector'
-import { MirroredMetric } from '../dataHub/metrics/mirrored'
-import { updateHistory } from '../../utils/utils'
-import styles from './index.module.scss'
+import styles from '../Studio2/index.module.scss'
+import { DEFAULT_SETTINGS } from '../Studio/defaults'
+import { saveToggle } from '../../utils/localStorage'
+import { mergeMetricSettingMap } from './utils'
 
-const Studio = ({
-  classes,
-  defaultSettings,
-  defaultOptions,
-  defaultMetrics,
-  defaultEvents,
-  defaultComparedMetrics,
-  defaultComparables,
-  defaultMetricSettingsMap,
-  topSlot,
-  bottomSlot,
-  ...props
+export const Studio = ({
+  defaultWidgets,
+  defaultSidepanel,
+  defaultSettings = DEFAULT_SETTINGS,
+  extensions,
 }) => {
-  const [project, setProject] = useState()
+  const [widgets, setWidgets] = useState(defaultWidgets)
   const [settings, setSettings] = useState(defaultSettings)
-  const [options, setOptions] = useState(defaultOptions)
-  const [comparables, setComparables] = useState(defaultComparables)
-  const [comparedMetrics, setComparedMetrics] = useState(defaultComparedMetrics)
-  const [metrics, setMetrics] = useState(defaultMetrics)
-  const [activeMetrics, setActiveMetrics] = useState(defaultMetrics)
-  const [activeEvents, setActiveEvents] = useState(defaultEvents)
-  const [MetricSettingMap, setMetricSettingMap] = useState(
-    defaultMetricSettingsMap
+  const [sidepanel, setSidepanel] = useState(defaultSidepanel)
+  const [selectedMetrics, setSelectedMetrics] = useState([])
+  const [selectedMetricSettingsMap, setSelectedMetricSettingsMap] = useState(
+    new Map(),
   )
-  const [chartSidepane, setChartSidepane] = useState()
-  const [advancedView, setAdvancedView] = useState()
-  const [shareLink, setShareLink] = useState()
+  const [isAnomalyActive, setIsAnomalyActive] = useState(false)
   const [isICOPriceDisabled, setIsICOPriceDisabled] = useState(true)
-  const [isSidebarClosed, setIsSidebarClosed] = useState()
-  const [MetricTransformer, setMetricTransformer] = useState({})
-  const [rawData, loadings, ErrorMsg] = useTimeseries(
-    activeMetrics,
-    settings,
-    MetricSettingMap,
-    MetricTransformer
-  )
-  const [eventsData, eventLoadings] = useTimeseries(activeEvents, settings)
-  const data = useClosestValueData(
-    rawData,
-    activeMetrics,
-    options.isClosestDataActive
-  )
-
-  useEffect(
-    () => {
-      setComparedMetrics(comparables.map(buildComparedMetric))
-    },
-    [comparables]
-  )
-
-  useEffect(
-    () => {
-      setActiveMetrics(metrics.concat(comparedMetrics))
-    },
-    [metrics, comparedMetrics]
-  )
-
-  useEffect(
-    () => {
-      const metricTransformer = Object.assign({}, MetricTransformer)
-      const lookupMetrics = options.isMultiChartsActive ? [] : metrics
-
-      metrics.forEach(metric => {
-        const mirrorOf = MirroredMetric[metric.key]
-        if (mirrorOf) {
-          const { key, preTransformer } = metric
-
-          if (lookupMetrics.includes(mirrorOf)) {
-            metricTransformer[key] = preTransformer
-          } else {
-            metricTransformer[key] = undefined
-          }
-        }
-      })
-
-      setMetricTransformer(metricTransformer)
-    },
-    [metrics, options.isMultiChartsActive]
-  )
-
-  useEffect(
-    () => {
-      if (Object.keys(ErrorMsg).length > 0) {
-        setMetrics(metrics.filter(({ key }) => !ErrorMsg[key]))
-      }
-    },
-    [ErrorMsg]
-  )
-
-  useEffect(
-    () => {
-      const activeLength = activeMetrics.length
-      if (!options.isMultiChartsActive && activeLength > MAX_METRICS_AMOUNT) {
-        const diff = activeLength - MAX_METRICS_AMOUNT
-
-        if (diff >= comparables.length) {
-          setComparables([])
-        } else {
-          setComparables(comparables.slice(0, -diff))
-        }
-
-        if (metrics.length >= MAX_METRICS_AMOUNT) {
-          setMetrics(metrics.slice(0, MAX_METRICS_AMOUNT))
-        }
-      }
-    },
-    [options.isMultiChartsActive]
-  )
+  const [isICOPriceActive, setIsICOPriceActive] = useState(true)
 
   useEffect(
     () => {
       const { slug } = defaultSettings
       if (slug && slug !== settings.slug) {
-        setSettings(state => ({ ...state, slug }))
+        setSettings({ ...settings, slug })
       }
     },
-    [defaultSettings.slug]
+    [defaultSettings.slug],
   )
 
-  useEffect(
-    () => {
-      const queryString =
-        '?' +
-        generateShareLink(settings, options, metrics, activeEvents, comparables)
+  function rerenderWidgets() {
+    setWidgets(widgets.slice())
+  }
 
-      const { origin, pathname } = window.location
-      setShareLink(origin + pathname + queryString)
-      updateHistory(queryString)
-    },
-    [settings, options, metrics, activeEvents, comparables]
-  )
+  function toggleSidepanel(key) {
+    setSidepanel(sidepanel === key ? undefined : key)
+  }
 
-  useEffect(
-    () => {
-      if (options.isAnomalyActive) {
-        setActiveEvents(buildAnomalies(metrics))
-      } else {
-        setActiveEvents([])
-      }
-    },
-    [metrics, options.isAnomalyActive]
-  )
+  function deleteWidget(widget) {
+    setWidgets(widgets.filter((w) => w !== widget))
+  }
 
-  useEffect(
-    () => {
-      if (chartSidepane === TOP_HOLDERS_PANE && options.isMultiChartsActive) {
-        setChartSidepane()
-      }
-    },
-    [chartSidepane, settings.slug, options.isMultiChartsActive]
-  )
+  function toggleWidgetMetric(widget, metric) {
+    const metrics = deduceMetrics(widget.metrics, metric)
 
-  function toggleMetric (metric) {
-    if (metric.comparedTicker) {
-      return removeComparedMetric(metric)
-    }
-
-    const metricSet = new Set(metrics)
-    if (metricSet.has(metric)) {
-      if (activeMetrics.length === 1) return
-      metricSet.delete(metric)
+    if (metrics.length === 0) {
+      deleteWidget(widget)
     } else {
-      if (
-        !options.isMultiChartsActive &&
-        activeMetrics.length === MAX_METRICS_AMOUNT
-      ) {
-        return
-      }
-
-      metricSet.add(metric)
-      trackMetricState(metric, true)
+      widget.metrics = metrics
+      setWidgets([...widgets])
     }
-    setMetrics([...metricSet])
   }
 
-  function toggleChartSidepane (sidepane) {
-    setChartSidepane(chartSidepane === sidepane ? undefined : sidepane)
+  function toggleAnomaly() {
+    setIsAnomalyActive(saveToggle('isAnomalyActive', !isAnomalyActive))
   }
 
-  function toggleAdvancedView (mode) {
-    setAdvancedView(advancedView === mode ? undefined : mode)
+  function toggleSelectionMetric(metric) {
+    setSelectedMetrics(deduceMetrics(selectedMetrics, metric))
   }
 
-  function removeComparedMetric ({ key }) {
-    setComparables(comparables.filter(comp => comp.key !== key))
+  function deduceMetrics(metrics, metric) {
+    const newMetrics = new Set(metrics)
+
+    if (newMetrics.has(metric)) {
+      newMetrics.delete(metric)
+    } else {
+      newMetrics.add(metric)
+    }
+
+    return [...newMetrics]
   }
 
-  function changeTimePeriod (from, to, timeRange) {
+  function changeTimePeriod(from, to, timeRange) {
     const interval = getNewInterval(from, to)
 
-    setSettings(state => ({
+    setSettings((state) => ({
       ...state,
       timeRange,
       interval: INTERVAL_ALIAS[interval] || interval,
       from: from.toISOString(),
-      to: to.toISOString()
+      to: to.toISOString(),
     }))
   }
 
+  function onSidebarItemClick(item) {
+    const { type, key } = item
+
+    if (type === 'sidepanel') {
+      toggleSidepanel(key)
+    } else if (type === 'ico_price') {
+      setIsICOPriceActive(!isICOPriceActive)
+    } else if (type === 'widget') {
+      if (key === 'holder_distribution') {
+        setWidgets([
+          ...widgets,
+          newHolderDistributionWidget({
+            scrollIntoViewOnMount: true,
+          }),
+        ])
+      }
+    } else {
+      toggleSelectionMetric(item)
+    }
+  }
+
+  function onWidgetClick(widget) {
+    const newMetrics = new Set([...widget.metrics, ...selectedMetrics])
+
+    widget.metrics = [...newMetrics]
+    widget.MetricSettingMap = mergeMetricSettingMap(
+      widget.MetricSettingMap,
+      selectedMetricSettingsMap,
+    )
+
+    rerenderWidgets()
+  }
+
+  function onNewChartClick() {
+    setWidgets([
+      ...widgets,
+      newChartWidget({
+        metrics: selectedMetrics,
+        MetricSettingMap: selectedMetricSettingsMap,
+      }),
+    ])
+  }
+
+  function onOverviewClose() {
+    setSelectedMetrics([])
+    setSelectedMetricSettingsMap(new Map())
+  }
+
   return (
-    <div
-      className={cx(
-        styles.wrapper,
-        classes.wrapper,
-        isSidebarClosed && styles.wrapper_wide
-      )}
-    >
-      <StudioSidebar
+    <div className={styles.wrapper}>
+      <Sidebar
         slug={settings.slug}
-        options={options}
-        activeMetrics={activeMetrics}
-        activeEvents={activeEvents}
-        advancedView={advancedView}
-        chartSidepane={chartSidepane}
-        ErrorMsg={ErrorMsg}
-        setOptions={setOptions}
-        toggleMetric={toggleMetric}
-        toggleAdvancedView={toggleAdvancedView}
-        toggleChartSidepane={toggleChartSidepane}
-        setIsSidebarClosed={setIsSidebarClosed}
-        setMetricSettingMap={setMetricSettingMap}
+        activeMetrics={selectedMetrics}
+        isAnomalyActive={isAnomalyActive}
         isICOPriceDisabled={isICOPriceDisabled}
-        isSidebarClosed={isSidebarClosed}
-        project={project}
+        toggleMetric={onSidebarItemClick}
+        toggleAnomaly={toggleAnomaly}
+        setMetricSettingMap={setSelectedMetricSettingsMap}
       />
-      <StudioMain
-        {...props}
-        onProjectChange={setProject}
-        topSlot={topSlot}
-        bottomSlot={bottomSlot}
-        options={options}
-        settings={settings}
-        activeMetrics={activeMetrics}
-        activeEvents={activeEvents}
-        metrics={metrics}
-        comparables={comparables}
-        data={data}
-        eventsData={eventsData}
-        loadings={loadings}
-        eventLoadings={eventLoadings}
-        advancedView={advancedView}
-        chartSidepane={chartSidepane}
-        shareLink={shareLink}
-        // bools
-        isSidebarClosed={isSidebarClosed}
-        // state setters
-        setMetrics={setMetrics}
-        setSettings={setSettings}
-        setOptions={setOptions}
-        setComparables={setComparables}
-        setIsICOPriceDisabled={setIsICOPriceDisabled}
-        // fn
-        toggleMetric={toggleMetric}
-        toggleAdvancedView={toggleAdvancedView}
-        toggleChartSidepane={toggleChartSidepane}
-        changeTimePeriod={changeTimePeriod}
-        project={project}
-      />
+      <main className={styles.main}>
+        <Main
+          widgets={widgets}
+          settings={settings}
+          sidepanel={sidepanel}
+          isICOPriceActive={isICOPriceActive}
+          isAnomalyActive={isAnomalyActive}
+          // fn
+          setWidgets={setWidgets}
+          setSettings={setSettings}
+          setIsICOPriceDisabled={setIsICOPriceDisabled}
+          changeTimePeriod={changeTimePeriod}
+          toggleWidgetMetric={toggleWidgetMetric}
+          toggleSidepanel={toggleSidepanel}
+          deleteWidget={deleteWidget}
+          rerenderWidgets={rerenderWidgets}
+        />
+
+        {selectedMetrics.length ? (
+          <SelectionOverview
+            widgets={widgets}
+            selectedMetrics={selectedMetrics}
+            toggleMetric={toggleSelectionMetric}
+            onClose={onOverviewClose}
+            onWidgetClick={onWidgetClick}
+            onNewChartClick={onNewChartClick}
+          />
+        ) : null}
+      </main>
+      {React.Children.map(extensions, (extension) =>
+        React.cloneElement(extension, { widgets, settings, sidepanel }),
+      )}
     </div>
   )
 }
 
-Studio.defaultProps = {
-  defaultComparedMetrics: [],
-  defaultEvents: [],
-  defaultComparables: [],
-  onSlugChange: () => {},
-  classes: {}
-}
-
-export default ({
-  settings,
-  options,
-  metrics,
-  events,
-  comparables,
-  MetricSettingsMap,
-  ...props
-}) => (
-  <Studio
-    {...props}
-    defaultSettings={{
-      ...DEFAULT_SETTINGS,
-      ...settings
-    }}
-    defaultOptions={{ ...DEFAULT_OPTIONS, ...options }}
-    defaultMetrics={metrics || DEFAULT_METRICS}
-    defaultEvents={events}
-    defaultComparables={comparables}
-    defaultMetricSettingsMap={MetricSettingsMap || DEFAULT_METRIC_SETTINGS_MAP}
-  />
-)
+export default Studio
