+++ src/ducks/Studio/utils.js
@@ -1,6 +1,6 @@
 import { MirroredMetric } from '../dataHub/metrics/mirrored'
 
-export function extractMirrorMetricsDomainGroups (domainGroups) {
+export function extractMirrorMetricsDomainGroups(domainGroups) {
   if (!domainGroups) return
 
   const mirroredGroups = []
@@ -19,4 +19,14 @@ export function extractMirrorMetricsDomainGroups (domainGroups) {
   }
 
   return mirroredGroups
+}
+
+export function mergeMetricSettingMap(oldMap, newMap) {
+  const mergedMap = new Map()
+
+  newMap.forEach((newSettings, metric) =>
+    mergedMap.set(metric, Object.assign({}, oldMap.get(metric), newSettings)),
+  )
+
+  return mergedMap
 }
