+++ src/ducks/Studio2/Tabs/MetricExplanation.js
@@ -1,12 +1,12 @@
 import React from 'react'
-import Sidepane from '../../Studio/Chart/Sidepane'
+import Sidepanel from '../../Studio/Chart/Sidepanel'
 import styles from '../index.module.scss'
 
 const MetricColor = {}
 
 const MetricExplanation = ({ widgets, toggleMetricExplanationVisibility }) => {
   return (
-    <Sidepane
+    <Sidepanel
       className={styles.side}
       chartSidepane='METRICS_EXPLANATION_PANE'
       metrics={widgets.reduce((acc, widget) => [...acc, ...widget.metrics], [])}
