--- src/ducks/Studio2/utils.js
@@ -1,9 +1,0 @@
-export function mergeMetricSettingMap(oldMap, newMap) {
-  const mergedMap = new Map()
-
-  newMap.forEach((newSettings, metric) =>
-    mergedMap.set(metric, Object.assign({}, oldMap.get(metric), newSettings)),
-  )
-
-  return mergedMap
-}
