+++ src/pages/Studio/URLExtension.js
@@ -1,5 +1,5 @@
 import { useEffect } from 'react'
-import { generateUrlV2 } from '../../ducks/Studio2/url/generate'
+import { generateUrlV2 } from '../../ducks/Studio/url/generate'
 import { updateHistory } from '../../utils/utils'
 
 const URLExtension = ({ history, settings, widgets, sidepanel }) => {
